Improve API structure and docs (#373)
* chore: move non-auth specific routes out of auth module, add login api docs

* chore(cleanup): move host, invites and user specific api routes into their own submodules

* chore(docs): add docs for /register api method, fix docs for /login

* chore(docs): add auth module docs

* chore(docs): add invite route docs

* chore(docs): add docs for `/api/v1/user` routes
