+++ dim/src/routes/auth.rs
@@ -1,10 +1,26 @@
+//! This module contains all docs and APIs related to authentication and user creation.
+//!
+//! # Request Authentication and Authorization
+//! Most API endpoints require a valid JWT authentication token. If no such token is supplied, the
+//! API will return [`Unauthenticated`]. Authentication tokens can be obtained by logging in with
+//! the [`login`] method. Authentication tokens must be passed to the server through a
+//! `Authroization` header.
+//!
+//! ## Example of an authenticated call
+//! ```text
+//! curl -X POST http://127.0.0.1:8000/api/v1/auth/whoami -H "Content-type: application/json" -H
+//! "Authorization: eyJhb....."
+//! ```
+//!
+//! # Token expiration
+//! By default tokens expire after exactly two weeks, once the tokens expire the client must renew
+//! them. At the moment renewing the token is only possible by logging in again.
+//!
+//! [`Unauthenticated`]: crate::errors::DimError::Unauthenticated
+//! [`login`]: fn@login
 use crate::core::DbConnection;
 use crate::errors;
-use bytes::BufMut;
 
-use database::asset::Asset;
-use database::asset::InsertableAsset;
-use database::progress::Progress;
 use database::user::verify;
 use database::user::InsertableUser;
 use database::user::Login;
@@ -14,21 +30,14 @@ use serde_json::json;
 
 use warp::reply;
 
-use http::StatusCode;
-
-use futures::TryStreamExt;
-use uuid::Uuid;
-
 pub mod filters {
     use crate::core::DbConnection;
-    use serde::Deserialize;
 
     use warp::reject;
     use warp::Filter;
 
     use database::user::Login;
 
-    use super::super::global_filters::with_auth;
     use super::super::global_filters::with_db;
 
     pub fn login(
@@ -43,22 +52,6 @@ pub mod filters {
                     .await
                     .map_err(|e| reject::custom(e))
             })
-    }
-
-    pub fn whoami(
-        conn: DbConnection,
-    ) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
-        warp::path!("api" / "v1" / "auth" / "whoami")
-            .and(warp::get())
-            .and(with_auth(conn.clone()))
-            .and(with_db(conn))
-            .and_then(
-                |auth: database::user::User, conn: DbConnection| async move {
-                    super::whoami(auth, conn)
-                        .await
-                        .map_err(|e| reject::custom(e))
-                },
-            )
     }
 
     pub fn admin_exists(
@@ -87,144 +80,35 @@ pub mod filters {
                     .map_err(|e| reject::custom(e))
             })
     }
-
-    pub fn get_all_invites(
-        conn: DbConnection,
-    ) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
-        warp::path!("api" / "v1" / "auth" / "invites")
-            .and(warp::get())
-            .and(with_auth(conn.clone()))
-            .and(with_db(conn))
-            .and_then(
-                |user: database::user::User, conn: DbConnection| async move {
-                    super::get_all_invites(conn, user)
-                        .await
-                        .map_err(|e| reject::custom(e))
-                },
-            )
-    }
-
-    pub fn generate_invite(
-        conn: DbConnection,
-    ) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
-        warp::path!("api" / "v1" / "auth" / "new_invite")
-            .and(warp::post())
-            .and(with_auth(conn.clone()))
-            .and(with_db(conn))
-            .and_then(
-                |user: database::user::User, conn: DbConnection| async move {
-                    super::generate_invite(conn, user)
-                        .await
-                        .map_err(|e| reject::custom(e))
-                },
-            )
-    }
-
-    pub fn user_change_password(
-        conn: DbConnection,
-    ) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
-        #[derive(Deserialize)]
-        pub struct Params {
-            old_password: String,
-            new_password: String,
-        }
-
-        warp::path!("api" / "v1" / "auth" / "password")
-            .and(warp::patch())
-            .and(with_auth(conn.clone()))
-            .and(warp::body::json::<Params>())
-            .and(with_db(conn))
-            .and_then(
-                |user: database::user::User,
-                 Params {
-                     old_password,
-                     new_password,
-                 }: Params,
-                 conn: DbConnection| async move {
-                    super::user_change_password(conn, user, old_password, new_password)
-                        .await
-                        .map_err(|e| reject::custom(e))
-                },
-            )
-    }
-
-    pub fn admin_delete_token(
-        conn: DbConnection,
-    ) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
-        warp::path!("api" / "v1" / "auth" / "token" / String)
-            .and(warp::delete())
-            .and(with_auth(conn.clone()))
-            .and(with_db(conn))
-            .and_then(
-                |token: String, auth: database::user::User, conn: DbConnection| async move {
-                    super::delete_invite(conn, auth, token)
-                        .await
-                        .map_err(|e| reject::custom(e))
-                },
-            )
-    }
-
-    pub fn user_delete_self(
-        conn: DbConnection,
-    ) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
-        #[derive(Deserialize)]
-        pub struct Params {
-            password: String,
-        }
-
-        warp::path!("api" / "v1" / "user" / "delete")
-            .and(warp::delete())
-            .and(with_auth(conn.clone()))
-            .and(warp::body::json::<Params>())
-            .and(with_db(conn))
-            .and_then(
-                |auth: database::user::User, Params { password }: Params, conn: DbConnection| async move {
-                    super::user_delete_self(conn, auth, password)
-                        .await
-                        .map_err(|e| reject::custom(e))
-                },
-            )
-    }
-
-    pub fn user_change_username(
-        conn: DbConnection,
-    ) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
-        #[derive(Deserialize)]
-        pub struct Params {
-            new_username: String,
-        }
-        warp::path!("api" / "v1" / "auth" / "username")
-            .and(warp::patch())
-            .and(with_auth(conn.clone()))
-            .and(warp::body::json::<Params>())
-            .and(with_db(conn))
-            .and_then(
-                |user: database::user::User,
-                 Params { new_username }: Params,
-                 conn: DbConnection| async move {
-                    super::user_change_username(conn, user, new_username)
-                        .await
-                        .map_err(|e| reject::custom(e))
-                },
-            )
-    }
-
-    pub fn user_upload_avatar(
-        conn: DbConnection,
-    ) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
-        warp::path!("api" / "v1" / "user" / "avatar")
-            .and(warp::post())
-            .and(with_auth(conn.clone()))
-            .and(warp::multipart::form().max_length(5_000_000))
-            .and(with_db(conn))
-            .and_then(|user, form, conn| async move {
-                super::user_upload_avatar(conn, user, form)
-                    .await
-                    .map_err(|e| reject::custom(e))
-            })
-    }
 }
 
+/// # POST `/api/v1/auth/login`
+/// Method will log a user in and return a authentication token that can be used to authenticate other
+/// requests.
+///
+/// # Request
+/// This method accepts a JSON body that deserializes into [`Login`].
+///
+/// ## Example
+/// ```text
+/// curl -X POST http://127.0.0.1:8000/api/v1/auth/login -H "Content-type: application/json" -d
+/// '{"username": "testuser", "password": "testpassword"}'
+/// ```
+///
+/// # Response
+/// If authentication is successful, this method will return status `200 0K` as well as a
+/// authentication token.
+/// ```
+/// {
+///   "token": "...."
+/// }
+/// ```
+///
+/// # Errors
+/// * [`InvalidCredentials`] - The provided username or password is incorrect.
+///
+/// [`InvalidCredentials`]: crate::errors::DimError::InvalidCredentials
+/// [`Login`]: database::user::Login
 pub async fn login(
     new_login: Login,
     conn: DbConnection,
@@ -245,19 +129,6 @@ pub async fn login(
     Err(errors::DimError::InvalidCredentials)
 }
 
-pub async fn whoami(user: User, conn: DbConnection) -> Result<impl warp::Reply, errors::DimError> {
-    let mut tx = conn.read().begin().await?;
-
-    Ok(reply::json(&json!({
-        "picture": Asset::get_of_user(&mut tx, user.id).await.ok().map(|x| format!("/images/{}", x.local_path)),
-        "spentWatching": Progress::get_total_time_spent_watching(&mut tx, user.id)
-            .await
-            .unwrap_or(0) / 3600,
-        "username": user.username,
-        "roles": user.roles()
-    })))
-}
-
 pub async fn admin_exists(conn: DbConnection) -> Result<impl warp::Reply, errors::DimError> {
     let mut tx = conn.read().begin().await?;
     Ok(reply::json(&json!({
@@ -265,6 +136,40 @@ pub async fn admin_exists(conn: DbConnection) -> Result<impl warp::Reply, errors
     })))
 }
 
+/// # POST `/api/v1/auth/register`
+/// Method will create a new user and return it a authentication token if a user has been
+/// successfuly created.
+///
+/// # Request
+/// This method accepts a JSON body that deserializes into [`Login`]. If there are no other users
+/// in the database, this route will give the new user `owner` permissions. Additionally this route
+/// will not require an invite token.
+///
+/// If there is a user in the database, this request will require an invite token and the user will
+/// be given only `user` permissions.
+///
+/// ## Example
+/// ```text
+/// curl -X POST http://127.0.0.1:8000/api/v1/auth/login -H "Content-type: application/json" -d
+/// '{"username": "testuser", "password": "testpassword", "invite_token":
+/// "72390330-b8af-4413-8305-5f8cae1c8f88"}'
+/// ```
+///
+/// # Response
+/// If a user is successfully created, this method will return status `200 0K` as well as the
+/// create user's username.
+/// ```
+/// {
+///   "username": "...."
+/// }
+/// ```
+///
+/// # Errors
+/// * [`NoToken`] - Either the request doesnt contain an invite token, or the invite token is
+/// invalid.
+///
+/// [`NoToken`]: crate::errors::DimError::NoToken
+/// [`Login`]: database::user::Login
 pub async fn register(
     new_user: Login,
     conn: DbConnection,
@@ -276,8 +181,7 @@ pub async fn register(
     let users_empty = User::get_all(&mut tx).await?.is_empty();
 
     if !users_empty
-        && (new_user.invite_token.is_none()
-            || !new_user.invite_token_valid(&mut tx).await.unwrap_or(false))
+        && (new_user.invite_token.is_none() || !new_user.invite_token_valid(&mut tx).await?)
     {
         return Err(errors::DimError::NoToken);
     }
@@ -309,202 +213,4 @@ pub async fn register(
     tx.commit().await?;
 
     Ok(reply::json(&json!({ "username": res.username })))
-}
-
-pub async fn get_all_invites(
-    conn: DbConnection,
-    user: User,
-) -> Result<impl warp::Reply, errors::DimError> {
-    let mut tx = conn.read().begin().await?;
-    if user.has_role("owner") {
-        #[derive(serde::Serialize)]
-        struct Row {
-            id: String,
-            created: i64,
-            claimed_by: Option<String>,
-        }
-
-        // FIXME: LEFT JOINs cause sqlx::query! to panic, thus we must get tokens in two queries.
-        let mut row = sqlx::query_as!(
-            Row,
-            r#"SELECT invites.id, invites.date_added as created, NULL as "claimed_by: _"
-                FROM invites
-                WHERE invites.id NOT IN (SELECT users.claimed_invite FROM users)
-                ORDER BY created ASC"#
-        )
-        .fetch_all(&mut tx)
-        .await
-        .unwrap_or_default();
-
-        row.append(
-            &mut sqlx::query_as!(
-                Row,
-                r#"SELECT invites.id, invites.date_added as created, users.username as "claimed_by: Option<String>"
-            FROM  invites
-            INNER JOIN users ON users.claimed_invite = invites.id"#
-            )
-            .fetch_all(&mut tx)
-            .await
-            .unwrap_or_default(),
-        );
-
-        return Ok(reply::json(&row));
-    }
-
-    Err(errors::DimError::Unauthorized)
-}
-
-pub async fn generate_invite(
-    conn: DbConnection,
-    user: User,
-) -> Result<impl warp::Reply, errors::DimError> {
-    if !user.has_role("owner") {
-        return Err(errors::DimError::Unauthorized);
-    }
-
-    let mut lock = conn.writer().lock_owned().await;
-    let mut tx = database::write_tx(&mut lock).await?;
-
-    let token = Login::new_invite(&mut tx).await?;
-
-    tx.commit().await?;
-
-    Ok(reply::json(&json!({ "token": token })))
-}
-
-pub async fn delete_invite(
-    conn: DbConnection,
-    user: User,
-    token: String,
-) -> Result<impl warp::Reply, errors::DimError> {
-    if !user.has_role("owner") {
-        return Err(errors::DimError::Unauthorized);
-    }
-
-    let mut lock = conn.writer().lock_owned().await;
-    let mut tx = database::write_tx(&mut lock).await?;
-    Login::delete_token(&mut tx, token).await?;
-    tx.commit().await?;
-
-    Ok(StatusCode::OK)
-}
-
-pub async fn user_change_password(
-    conn: DbConnection,
-    user: User,
-    old_password: String,
-    new_password: String,
-) -> Result<impl warp::Reply, errors::DimError> {
-    let mut lock = conn.writer().lock_owned().await;
-    let mut tx = database::write_tx(&mut lock).await?;
-    let user = User::authenticate(&mut tx, user.username, old_password)
-        .await
-        .map_err(|_| errors::DimError::InvalidCredentials)?;
-    user.set_password(&mut tx, new_password).await?;
-
-    tx.commit().await?;
-
-    Ok(StatusCode::OK)
-}
-
-pub async fn user_delete_self(
-    conn: DbConnection,
-    user: User,
-    password: String,
-) -> Result<impl warp::Reply, errors::DimError> {
-    let mut lock = conn.writer().lock_owned().await;
-    let mut tx = database::write_tx(&mut lock).await?;
-    let _ = User::authenticate(&mut tx, user.username, password)
-        .await
-        .map_err(|_| errors::DimError::InvalidCredentials)?;
-
-    User::delete(&mut tx, user.id).await?;
-
-    tx.commit().await?;
-
-    Ok(StatusCode::OK)
-}
-
-pub async fn user_change_username(
-    conn: DbConnection,
-    user: User,
-    new_username: String,
-) -> Result<impl warp::Reply, errors::DimError> {
-    let mut lock = conn.writer().lock_owned().await;
-    let mut tx = database::write_tx(&mut lock).await?;
-    if User::get(&mut tx, &new_username).await.is_ok() {
-        return Err(errors::DimError::UsernameNotAvailable);
-    }
-
-    User::set_username(&mut tx, user.username, new_username).await?;
-    tx.commit().await?;
-
-    Ok(StatusCode::OK)
-}
-
-pub async fn user_upload_avatar(
-    conn: DbConnection,
-    user: User,
-    form: warp::multipart::FormData,
-) -> Result<impl warp::Reply, errors::DimError> {
-    let parts: Vec<warp::multipart::Part> = form
-        .try_collect()
-        .await
-        .map_err(|_e| errors::DimError::UploadFailed)?;
-
-    let mut lock = conn.writer().lock_owned().await;
-    let mut tx = database::write_tx(&mut lock).await?;
-    let asset = if let Some(p) = parts.into_iter().filter(|x| x.name() == "file").next() {
-        process_part(&mut tx, p).await
-    } else {
-        Err(errors::DimError::UploadFailed)
-    };
-
-    User::set_picture(&mut tx, user.id, asset?.id).await?;
-    tx.commit().await?;
-
-    Ok(StatusCode::OK)
-}
-
-pub async fn process_part(
-    conn: &mut database::Transaction<'_>,
-    p: warp::multipart::Part,
-) -> Result<Asset, errors::DimError> {
-    if p.name() != "file" {
-        return Err(errors::DimError::UploadFailed);
-    }
-
-    let file_ext = match dbg!(p.content_type()) {
-        Some("image/jpeg" | "image/jpg") => "jpg",
-        Some("image/png") => "png",
-        _ => return Err(errors::DimError::UnsupportedFile),
-    };
-
-    let contents = p
-        .stream()
-        .try_fold(Vec::new(), |mut vec, data| {
-            vec.put(data);
-            async move { Ok(vec) }
-        })
-        .await
-        .map_err(|_| errors::DimError::UploadFailed)?;
-
-    let local_file = format!("{}.{}", Uuid::new_v4().to_string(), file_ext);
-    let local_path = format!(
-        "{}/{}",
-        crate::core::METADATA_PATH.get().unwrap(),
-        &local_file
-    );
-
-    tokio::fs::write(&local_path, contents)
-        .await
-        .map_err(|_| errors::DimError::UploadFailed)?;
-
-    Ok(InsertableAsset {
-        local_path: local_file,
-        file_ext: file_ext.into(),
-        ..Default::default()
-    }
-    .insert(conn)
-    .await?)
 }
